refactor: isolate RemixAgent
